reformat

+++ s3_autoupload.py
@@ -13,13 +13,18 @@ class S3Uploader(FileSystemEventHandler):
         super().__init__()
         # create an s3 client and pass in the credentials, url and region
         self.client = boto3.client(
-            's3', aws_access_key_id=key, aws_secret_access_key=secret, endpoint_url=url, region_name='us-east-1')
+            's3',
+            aws_access_key_id=key,
+            aws_secret_access_key=secret,
+            endpoint_url=url,
+            region_name='us-east-1'
+        )
         self.bucket = bucket
         self.basepath = basepath
         print("ready to upload files to s3")
 
-    """compare all files in the directory to the files in the bucket and upload any new files, delete any files that are no longer in the directory"""
     def mirror(self):
+        """compare all files in the directory to the files in the bucket and upload any new files, delete any files that are no longer in the directory"""
         print("initial mirroring of files...")
         # get all the files in the directory
         local_files = set()
@@ -33,25 +38,31 @@ class S3Uploader(FileSystemEventHandler):
         # get all the files in the bucket
         file_in_bucket = set()
         response = self.client.list_objects_v2(Bucket=self.bucket)
-        if(response['KeyCount'] > 0):
+        if response['KeyCount'] > 0:
             for obj in response['Contents']:
                 file_in_bucket.add(obj['Key'])
+
             while response['IsTruncated']:
                 response = self.client.list_objects_v2(
                     Bucket=self.bucket, ContinuationToken=response['NextContinuationToken'])
+
                 for obj in response['Contents']:
                     file_in_bucket.add(obj['Key'])
 
         for obj in file_in_bucket:
             if obj not in local_files:
                 files_to_delete.add(obj)
+
             elif obj in files_to_upload:
                 files_to_upload.remove(obj)
 
         # upload any new files
         for file in files_to_upload:
             self.client.upload_file(
-                self.basepath.joinpath(file), self.bucket, file)
+                self.basepath.joinpath(file),
+                self.bucket,
+                file
+            )
             print(f"Uploaded {file} to {self.bucket}")
 
         # delete any files that are no longer in the directory
@@ -62,23 +73,21 @@ class S3Uploader(FileSystemEventHandler):
         print("initial mirroring complete")
 
     def on_created(self, event: FileSystemEvent) -> None:
-        try:
-            filepath = Path(event.src_path).relative_to(
-                self.basepath).as_posix()
-            self.client.upload_file(
-                event.src_path, self.bucket, filepath)
-            print(f"Uploaded {event.src_path} to {self.bucket}")
-        except Exception as e:
-            print(f"Error uploading {event.src_path} to {self.bucket}: {e}")
+        filepath = Path(event.src_path)
+        if not filepath.exists() or not filepath.is_file():
+            return
+
+        self.client.upload_file(
+            event.src_path,
+            self.bucket,
+            filepath.relative_to(self.basepath).as_posix()
+        )
+        print(f"Uploaded {event.src_path} to {self.bucket}")
 
     def on_deleted(self, event: FileSystemEvent) -> None:
-        try:
-            filepath = Path(event.src_path).relative_to(
-                self.basepath).as_posix()
-            self.client.delete_object(Bucket=self.bucket, Key=filepath)
-            print(f"Deleted {event.src_path} from {self.bucket}")
-        except Exception as e:
-            print(f"Error deleting {event.src_path} from {self.bucket}: {e}")
+        filepath = Path(event.src_path).relative_to(self.basepath).as_posix()
+        self.client.delete_object(Bucket=self.bucket, Key=filepath)
+        print(f"Deleted {event.src_path} from {self.bucket}")
 
     def on_modified(self, event: FileSystemEvent) -> None:
         self.on_created(event)
@@ -100,14 +109,15 @@ class Watcher:
         try:
             while True:
                 time.sleep(1)
+
         except KeyboardInterrupt:
             self.observer.stop()
+
         self.observer.join()
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(
-        description="Watch a directory for changes and upload to S3")
+    parser = argparse.ArgumentParser(description="Watch a directory for changes and upload to S3")
     parser.add_argument('dir', help="Directory to watch")
     parser.add_argument('bucket', help="S3 bucket name")
     parser.add_argument('url', help="S3 endpoint url")
@@ -117,7 +127,6 @@ if __name__ == "__main__":
 
     path = Path(args.dir)
 
-    # create an instance of the s3_uploader class
     s3handler = S3Uploader(args.bucket, args.url, args.key, args.secret, path)
     s3handler.mirror()
